Extract chatStore.ts WebSocket handlers into domain-specific modules
Break the 1307-line monolithic chatStore into a thin dispatcher (484 lines)
plus domain-specific handler and helper modules:

- handlers/streamingHandlers.ts — thinking, token, tool_use, tool_result, done, stopped, error
- handlers/sessionHandlers.ts — 8 session lifecycle handlers + answer_injected
- handlers/panelHandlers.ts — plan_update, subagent_start/complete, hoa_progress
- handlers/auxiliaryHandlers.ts — interaction, file_changed, notifications, background_tasks
- helpers/blockHelpers.ts — panel block append/update, auto-close timers
- helpers/bufferReplay.ts — session reconnect replay, deriveStatus, extractTodos

Eliminates all 18 `as any` casts via proper Extract<WSMessage, { type }> narrowing.
Deduplicates done/stopped panel cleanup into shared finalizeRunningPanels().
No component import changes — purely internal reorganization.
